Split MP REST Client integration test module to support javax.servlet in the TCK
