+++ tests/integration/microprofile/pom.xml
@@ -33,6 +33,7 @@
     <modules>
         <module>config</module>
         <module>rest-client</module>
+        <module>rest-client-tck</module>
     </modules>
 
     <build>
