+++ tests/integration/microprofile/rest-client/src/test/java/org/glassfish/jersey/test/microprofile/restclient/RestClientModelTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -14,30 +14,25 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.jersey.restclient;
+package org.glassfish.jersey.test.microprofile.restclient;
 
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.concurrent.CountDownLatch;
+import java.util.List;
+import java.util.Map;
 
-import org.apache.http.conn.ConnectionRequest;
-import org.apache.http.conn.routing.HttpRoute;
-import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
 import org.eclipse.microprofile.rest.client.RestClientBuilder;
-import org.glassfish.jersey.apache.connector.ApacheClientProperties;
-import org.glassfish.jersey.apache.connector.ApacheConnectorProvider;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.test.JerseyTest;
 import org.glassfish.jersey.test.TestProperties;
 import org.junit.Test;
 
-import static org.testng.Assert.assertEquals;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Created by David Kral.
  */
-public class ConnectorTest extends JerseyTest {
-
+public class RestClientModelTest extends JerseyTest {
     @Override
     protected ResourceConfig configure() {
         enable(TestProperties.LOG_TRAFFIC);
@@ -45,26 +40,20 @@ public class ConnectorTest extends JerseyTest {
     }
 
     @Test
-    public void testConnector() throws URISyntaxException {
-        CountDownLatch countDownLatch = new CountDownLatch(1);
-        PoolingHttpClientConnectionManager connectionManager = new PoolingHttpClientConnectionManager() {
-
-            @Override
-            public ConnectionRequest requestConnection(HttpRoute route, Object state) {
-                countDownLatch.countDown();
-                return super.requestConnection(route, state);
-            }
-
-        };
-
+    public void testGetIt() throws URISyntaxException {
         ApplicationResource app = RestClientBuilder.newBuilder()
                 .baseUri(new URI("http://localhost:9998"))
-                .property(ApacheClientProperties.CONNECTION_MANAGER, connectionManager)
-                .register(ApacheConnectorProvider.class)
                 .build(ApplicationResource.class);
+        List<String> collection = app.getValue();
+        assertEquals(2, collection.size());
+        assertEquals("This is default value!", collection.get(0));
+        assertEquals("Test", collection.get(1));
 
-        app.getTestMap();
-        assertEquals(countDownLatch.getCount(), 0);
+        Map<String, String> map = app.getTestMap();
+        assertEquals(2, map.size());
+        assertEquals("firstValue", map.get("firstKey"));
+        assertEquals("secondValue", map.get("secondKey"));
+
+        assertEquals("Hi", app.sayHi());
     }
-
 }
